refactor: WebChessGame을 제거
ChessGame Web에서도 사용할 수 있도록 메서드 추가
ChessWebController 리팩토링에 따른 수정
WebChessGame 삭제
